requirements.txt
fastapi
uvicorn[standard]
httpx
redis
python-dotenv
pydantic
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT
sqlalchemy
email-validator
pytest
pytest-asyncio
pytest-cov
pydantic-settings